requirements.txt
aiogram==3.13.0
asyncpg==0.30.0
asyncpg-lite==0.3.1.3
openpyxl==3.1.5
pandas==2.2.3
python-decouple==3.8
SQLAlchemy==2.0.36